requirements_advanced.txt
# Enhanced requirements for advanced features
SpeechRecognition==3.10.0
pyttsx3==2.90
pyautogui==0.9.54
pyperclip==1.8.2
pygetwindow==0.0.9
keyboard==0.13.5
psutil==5.9.5
screen-brightness-control==0.14.2
requests==2.31.0
pathlib==1.0.1
openai==1.3.7
google-generativeai==0.3.0
faster-whisper==0.10.0

# New dependencies for advanced features
Pillow==10.0.0
opencv-python==4.8.1.78
pystray==0.19.4
pywin32==306
numpy==1.24.3
schedule==1.2.0
plyer==2.1.0
rich==13.5.2
